Запуск и передача параметров в потоки перенесены в класс ThreatsIntro
